backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.2.1
SQLAlchemy>=2.0.30
greenlet>=3.0.3
asyncpg>=0.29.0
alembic>=1.13.1
redis>=5.0.4
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
structlog>=24.1.0
httpx>=0.27.0
email-validator>=2.1.1
reportlab>=4.2.0
jsonschema>=4.20.0
python-multipart>=0.0.9